implement validation when there are developers associated with qualification level

+++ src/modules/developers/useCases/deleteQualificationLevel/DeleteQualificationLevelUseCase.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from "tsyringe";
 import { AppError } from "../../../../shared/errors/AppError";
+import { DevelopersRepository } from "../../infra/typeorm/repositories/DevelopersRepository";
 import { IQualificationLevelsRepository } from "../../repositories/IQualificationLevelsRepository";
 
 
@@ -7,13 +8,20 @@ import { IQualificationLevelsRepository } from "../../repositories/IQualificatio
 class DeleteQualificationLevelUseCase {
     constructor(
         @inject("QualificationLevelsRepository")
-        private qualificationLevelsRepository: IQualificationLevelsRepository
+        private qualificationLevelsRepository: IQualificationLevelsRepository,
+        @inject("DevelopersRepository")
+        private developersRepository: DevelopersRepository
     ) {}
 
     async execute(id: number): Promise<void> {
         const qualificationLevel = await this.qualificationLevelsRepository.findById(id);
         if (!qualificationLevel) {
             throw new AppError("O nível informado não existe.")
+        }
+
+        const developers = await this.developersRepository.findByQualificationLevelId(qualificationLevel.id);
+        if (developers) {
+            throw new AppError("O nível não pode ser excluido, pois há desenvolvedores associados à ele.")
         }
 
         await this.qualificationLevelsRepository.delete(qualificationLevel.id);
